+++ Shopping List/Resources/Assets.xcassets/Apple.imageset/Contents.json
