+++ Shopping List/Resources/Assets.xcassets/Contents.json
