+++ Shopping List/Resources/Assets.xcassets/Milk.imageset/Contents.json
