+++ Shopping List/Resources/Assets.xcassets/Milk.imageset/milk.png
