+++ Shopping List/Resources/Assets.xcassets/Soda.imageset/soda.png
