+++ Shopping List/Resources/Assets.xcassets/Strawberries.imageset/Contents.json
